modified anomaly detection to work also on data drift, cleaned up data drift
