+++ models/schema.yml
@@ -1,58 +1,55 @@
 version: 2
 
 models:
-#   - name: boston
-#     description: Ozone levels
-#     config:
-#       materialized: table
-#     meta:
-#       owner: "@meder"
-#       fal:
-#         scripts:
-#           - fal_scripts/slack.py
-#           - fal_scripts/send_datadog_event.py
-#   - name: stg_zendesk_ticket_data
-#     description: zendesk ticket data
-#     config:
-#       materialized: table
-#     meta:
-#       owner: "@meder"
-#       fal:
-#         scripts:
-#           - fal_scripts/slack.py
-#           - fal_scripts/write_to_firestore.py
-#           - fal_scripts/zendesk_sentiment_analysis.py
-#   - name: zendesk_ticket_metrics
-#     description: Zendesk ticket metrics
-#     config:
-#       materialized: table
-#     meta:
-#       fal:
-#         owner: "@gorkem"
-#         scripts:
-#           - fal_scripts/forecast_slack.py
-#   - name: lombardia_covid
-#     description: Lombardia (IT) Covid19 Cases
-#     config:
-#       materialized: table
-#     meta:
-#       fal:
-#         owner: "@omer"
-#         scripts:
-#           - fal_scripts/anomaly_detection.py
+  - name: boston
+    description: Ozone levels
+    config:
+      materialized: table
+    meta:
+      owner: "@meder"
+      fal:
+        scripts:
+          - fal_scripts/slack.py
+          - fal_scripts/send_datadog_event.py
+  - name: stg_zendesk_ticket_data
+    description: zendesk ticket data
+    config:
+      materialized: table
+    meta:
+      owner: "@meder"
+      fal:
+        scripts:
+          - fal_scripts/slack.py
+          - fal_scripts/write_to_firestore.py
+          - fal_scripts/zendesk_sentiment_analysis.py
+  - name: zendesk_ticket_metrics
+    description: Zendesk ticket metrics
+    config:
+      materialized: table
+    meta:
+      fal:
+        owner: "@gorkem"
+        scripts:
+          - fal_scripts/forecast_slack.py
+  - name: lombardia_covid
+    description: Lombardia (IT) Covid19 Cases
+    config:
+      materialized: table
+    meta:
+      fal:
+        scripts:
+          - fal_scripts/anomaly_detection.py
   - name: city_temperatures
     description: New Orleans Temperatures
     config:
       materialized: table
     meta:
       fal:
-        owner: "@omer"
         scripts:
-          # - fal_scripts/data_drift.py
-          - fal_scripts/anomaly_detection.py
+          - fal_scripts/data_drift.py
 
-# fal:
-#   scripts:
-#     - fal_scripts/upload_to_gcs.py
-#     - fal_scripts/upload_to_s3.py
-#     - fal_scripts/list_owners_of_models.py
+fal:
+  scripts:
+    - fal_scripts/upload_to_gcs.py
+    - fal_scripts/upload_to_s3.py
+    - fal_scripts/list_owners_of_models.py
